Refactored key/mouse callbacks and changed to using glfw time (double) instead of System#nanoTime (long).

+++ src/main/java/supersnake/attribute/Updatable.java
@@ -2,5 +2,5 @@
 package supersnake.attribute;
 
 public interface Updatable{
-  public void update(long timeElapsed);
+  public void update(double timeElapsed);
 }
